LLM-Presentation-Generator/requirements.txt
fastapi
uvicorn
flask
dash
dash-bootstrap-components
pyngrok
python-pptx
torch
transformers
nest-asyncio
mangum
json-repair
